Move CellCycleAccuracy out of ctc submodule into stand alone submodule -- wip typing issue
