+++ src/traccuracy/metrics/__init__.py
@@ -1,6 +1,7 @@
 from ._basic import BasicMetrics
+from ._cca import CellCycleAccuracy
 from ._chota import CHOTAMetric
-from ._ctc import AOGMMetrics, CellCycleAccuracy, CTCMetrics
+from ._ctc import AOGMMetrics, CTCMetrics
 from ._divisions import DivisionMetrics
 from ._results import Results
 from ._track_overlap import TrackOverlapMetrics
